requirements.txt
aliyun-python-sdk-core==2.13.18
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-eci==1.0.9
aliyun-python-sdk-ecs==4.19.8
aliyun-python-sdk-kms==2.10.1
aliyun-python-sdk-vpc==3.0.9
amqp==2.6.1
anyjson
apache-libcloud==3.1.0
APScheduler==3.7.0
Babel==2.9.0
backcall==0.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
crcmod==1.7
decorator==4.4.2
Flask==1.1.2
Flask-RESTful==0.3.8
flower==0.9.5
future==0.18.2
futures==3.1.1
gevent==20.6.2
greenlet==0.4.16
humanize==3.1.0
idna==2.9
importlib-metadata==1.7.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
kafka-python==2.0.2
kombu==4.6.11
MarkupSafe==1.1.1
oss2==2.12.1
parso==0.7.0
Paste==3.4.3
peewee==3.13.3
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
Pygments==2.6.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-gitlab==1.4.0
python-hosts==1.0.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
rq==1.4.3
six==1.15.0
tornado==6.1
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.1.9
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
python-jenkins==1.7.0
